% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alt_rep.R
\name{alt_rep}
\alias{alt_rep}
\title{Compact replication}
\usage{
alt_rep(x, times)
}
\arguments{
\item{x}{scalar. Supports double, integer and bool}

\item{times}{an integer giving the umber of times to repeat x}
}
\description{
Compact replication using the ALTREP framework. It creates a sequence of n identical values but does
not materialize it unless it is necessary.
}
\examples{
x = alt_rep(2L, 1000L)
}
