% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alt_rep.R
\name{is_altrep}
\alias{is_altrep}
\alias{is_materialized}
\title{ALTREP tools}
\usage{
is_altrep(x)

is_materialized(x)
}
\arguments{
\item{x}{R object}
}
\description{
ALTREP tools
}
\examples{
u = alt_rep(2.1, 5)
is_altrep(u)
is_materialized(u)
}
